/dev/sysevent: correct type in event_format
VELAPLATFO-9514

jsonstr buffer size is calculated by get_long_integer_size,
%lld in specific platform might cause a long type value occupies more spaces than it real does.

+++ drivers/sysevent/sysevent.c
@@ -98,8 +98,8 @@ static int event_format(struct sysevent_s *event, char *jsonstr)
 {
   struct sysevent_payload_s *payload;
 
-  sprintf(jsonstr, "EventId %d -t %lld -paraList ", event->eventid,
-          event->ts.tv_sec);
+  sprintf(jsonstr, "EventId %d -t %ld -paraList ", event->eventid,
+          (long)event->ts.tv_sec);
   payload = event->head->next;
   if (payload == NULL)
     {
